requirements.txt
asgiref==3.7.2
blessed==1.20.0
boto3==1.28.76
botocore==1.31.76
cement==2.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.3
cpp_aws_s3_pdf==0.0.1
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
django-guardian==2.4.0
fontawesomefree==6.4.2
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.0
pathspec==0.10.1
pypdf==3.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
semantic-version==2.8.5
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
